Wrong data types for new widths and heights
As specified in bmp.h the width and  heights are longs and not integers. If a person is trying to resize a larger image having the width and height as a long is more useful

+++ pset4/resize/less/resize.c
@@ -57,10 +57,10 @@ int main(int argc, char *argv[])
     }
 
     // determine new dimensions
-    int oldWidth = bi.biWidth;
-    int oldHeight = bi.biHeight;
-    int newWidth = oldWidth * n;
-    int newHeight = oldHeight * n;
+    long oldWidth = bi.biWidth;
+    long oldHeight = bi.biHeight;
+    long newWidth = oldWidth * n;
+    long newHeight = oldHeight * n;
 
     // determine padding for scanlines
     int inPadding = (4 - (oldWidth * sizeof(RGBTRIPLE)) % 4) % 4;
